[native] Add personal thread icon
Summary: I think this icon makes sense - there are exactly two people on it

Test Plan: We're currently displaying thread icon in a child threads section of the settings and personal threads shouldn't be subthreads. Because of that, I changed a type of one of subthreads to PERSONAL and verified if the icon was displayed correctly.

Reviewers: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D329