[native] Display relationship prompt
Summary: A prompt is displayed only in personal threads

Test Plan: I verified that the prompt is displayed on in personal threads. I also created a couple of users, with every possible relationship and verified that the proper actions are displayed. Then I checked if the actions work correctly - I tested most of the and they worked fine. There is one issue I noticed, bun I still don't know how to fix: when a user blocks another user, a button which allows unblocking is wider than expected. It happens only for that single button and reopening the screen fixes the issue.

Reviewers: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D340