[native] Display a list of pending threads
Summary: When search text is not empty, we append pending threads to the end of the list. The pending thread is displayed only for users which do not have any relationship with viewer and are not together in a personal thread.

Test Plan: I searched for a user and pending thread was displayed at the end of the list. Then I removed search phrase and pending threads disappeared. Opening pending thread resulted in displaying the thread, without any messages and with relationship prompt. Sending friend request or a message caused a thread creation. Then the pending thread was no longer displayed.

Reviewers: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D373