[server] Do not create a new PERSONAL thread when it already exists
Test Plan: Sending a request to a user without a personal thread creates one. Sending a request to a user with a personal thread does not create a new one and only sends a message. The same applies for accepting a friend request. I also tested it in the more tab, where I tried to send a request to couple of users, from which couple had and couple hadn't personal thread.

Reviewers: ashoat

Subscribers: KatPo, zrebcu411, Adrian

Differential Revision: https://phabricator.ashoat.com/D397